Move project detection logic to a util
