+++ tool-plugins/vscode/package.json
@@ -212,8 +212,8 @@
                 "category": "Ballerina"
             },
             {
-                "command": "ballerina.runAllTests",
-                "title": "Run All Tests",
+                "command": "ballerina.project.test",
+                "title": "Run Project Tests",
                 "category": "Ballerina"
             },
             {
