+++ tool-plugins/vscode/src/telemetry/components.ts
@@ -4,6 +4,7 @@ export const CMP_DIAGRAM_VIEW = "component.diagram.view";
 export const CMP_DOCS_PREVIEW = "component.docs.preview";
 export const CMP_PROJECT_OVERVIEW = "component.project.overview";
 export const CMP_PROJECT_SUPPORT = "component.project.support";
-export const CMP_TEST_RUNNER = "component.test.runner";
+export const CMP_PROJECT_TEST_RUNNER = "component.project.test.runner";
+export const CMP_PROJECT_BUILD = "component.project.test.runner";
 export const CMP_TRACE_LOGS = "component.trace.logs";
 export const CMP_DEBUGGER = "component.debugger";
